requirements.txt
kafka-python==2.0.2
pyspark==3.5.0
openai==1.30.0
python-dotenv==1.0.0
pyyaml==6.0.1
pandas==2.1.0
scikit-learn==1.3.0
google-cloud-bigquery==3.13.0
apache-airflow==2.8.0
pytest==7.4.0
great-expectations==0.18.0